+++ tests/build-lib/project/lib/testlib/testmodule/Foo.cpp
@@ -1,5 +1,5 @@
 #include <testlib/testmodule/Foo.h>
 
 int getMagicNumber() {
-  return MAGIC_NUMBER; // MAGIC_NUMBER is defined via compiler flags
+  return MAGIC_NUMBER_A + MAGIC_NUMBER_B; // MAGIC_NUMBER_x is defined via compiler flags
 }